pyproject.toml
[tool.poetry]
name = "mtproto-mitm"
version = "0.1.0b1"
description = "Telegram mtproto mitm server."
authors = ["RuslanUC <dev_ruslan_uc@protonmail.com>"]
readme = "README.md"
packages = [{include = "mtproto_mitm"}]

[tool.poetry.dependencies]
python = "^3.11"
pysocks5server = "^0.1.0"
tgcrypto = "^1.2.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
